+++ src/test/java/lotto/domain/Lotto/LottoParserTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.Lotto;
 
 import lotto.domain.Lotto.LottoParser;
 import lotto.domain.LottoNumber.LottoNumber;
