+++ src/test/java/lotto/domain/Lotto/LottosTest.java
