+++ src/test/java/lotto/domain/LottoMoney/LottoMoneyTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.LottoMoney;
 
 import lotto.domain.LottoMoney.InvalidLottoMoneyException;
 import lotto.domain.LottoMoney.LottoMoney;
